Moved all files out of SchemaManager.Core and renamed namespaces.
