+++ Microsoft.Health.sln
@@ -61,8 +61,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Client", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Client.UnitTests", "src\Microsoft.Health.Client.UnitTests\Microsoft.Health.Client.UnitTests.csproj", "{D4DD763B-5246-47F4-A618-211AA820B65E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SchemaManager.Core", "tools\SchemaManager.Core\SchemaManager.Core.csproj", "{72816760-BDE2-4CAF-AF41-F7BEE8E26158}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SchemaManager.Core.UnitTests", "test\SchemaManager.Core.UnitTests\SchemaManager.Core.UnitTests.csproj", "{0BEB2EE3-DBC3-4DC4-BB01-141DBFFED8D2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.SqlServer.Tests.Integration", "test\Microsoft.Health.SqlServer.Tests.Integration\Microsoft.Health.SqlServer.Tests.Integration.csproj", "{5429C3A8-3429-43EA-BA05-E347CB4D1F61}"
@@ -185,10 +183,6 @@ Global
 		{D4DD763B-5246-47F4-A618-211AA820B65E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D4DD763B-5246-47F4-A618-211AA820B65E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D4DD763B-5246-47F4-A618-211AA820B65E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{72816760-BDE2-4CAF-AF41-F7BEE8E26158}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{72816760-BDE2-4CAF-AF41-F7BEE8E26158}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{72816760-BDE2-4CAF-AF41-F7BEE8E26158}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{72816760-BDE2-4CAF-AF41-F7BEE8E26158}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0BEB2EE3-DBC3-4DC4-BB01-141DBFFED8D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0BEB2EE3-DBC3-4DC4-BB01-141DBFFED8D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0BEB2EE3-DBC3-4DC4-BB01-141DBFFED8D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -272,7 +266,6 @@ Global
 		{83A62677-D82C-49A6-92E9-0D188C941A02} = {CCD9FF99-E177-446E-B9E5-9F570FD96A34}
 		{0EAE32B6-97C9-43C8-ABA5-8C0BAD9ED864} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 		{D4DD763B-5246-47F4-A618-211AA820B65E} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
-		{72816760-BDE2-4CAF-AF41-F7BEE8E26158} = {B70945F4-01A6-4351-955B-C4A2943B5E3B}
 		{0BEB2EE3-DBC3-4DC4-BB01-141DBFFED8D2} = {CCD9FF99-E177-446E-B9E5-9F570FD96A34}
 		{5429C3A8-3429-43EA-BA05-E347CB4D1F61} = {CCD9FF99-E177-446E-B9E5-9F570FD96A34}
 		{6D06503C-E79B-4F05-AFE1-C25386220895} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
@@ -289,7 +282,7 @@ Global
 		{9163ACB0-DF25-4924-A721-72BFAD9C430E} = {8AD2A324-DAB5-4380-94A5-31F7D817C384}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		RESX_SortFileContentOnSave = True
 		SolutionGuid = {562bbd07-1817-4af5-ab66-8837cf849248}
+		RESX_SortFileContentOnSave = True
 	EndGlobalSection
 EndGlobal
