+++ src/Microsoft.Health.SqlServer/Resources.resx
@@ -120,6 +120,9 @@
   <data name="AvailableVersionsDefaultErrorMessage" xml:space="preserve">
     <value>There are no available versions.</value>
   </data>
+  <data name="AvailableVersionsErrorMessage" xml:space="preserve">
+    <value>The available versions are not up-to-date.</value>
+  </data>
   <data name="BaseScriptNotFound" xml:space="preserve">
     <value>The base script is not found.</value>
   </data>
@@ -156,6 +159,10 @@
     <value>{0} is an invalid SQL identifier for a database.</value>
     <comment>{0} is the database identifier</comment>
   </data>
+  <data name="InvalidVersionMessage" xml:space="preserve">
+    <value>The schema version {0} cannot be applied because all the instances are not running the previous version.</value>
+    <comment>{0} is a version.</comment>
+  </data>
   <data name="OperationFailed" xml:space="preserve">
     <value>The SQL operation has failed.</value>
   </data>
@@ -166,8 +173,16 @@
   <data name="ScriptNotFound" xml:space="preserve">
     <value>The provided version is unknown.</value>
   </data>
+  <data name="SpecifiedVersionNotAvailable" xml:space="preserve">
+    <value>The specified schema version "{0}" is not available.</value>
+    <comment>{0} is a version.</comment>
+  </data>
   <data name="StringTooLong" xml:space="preserve">
     <value>Value of length {0} is too long for column {1} with length {2}</value>
     <comment>{0} is length of string data, {1} is the name of the column, {2} is the max allowable size of the column</comment>
+  </data>
+  <data name="VersionIncompatibilityMessage" xml:space="preserve">
+    <value>The schema version "{0}" is not compatible.</value>
+    <comment>{0} is a version.</comment>
   </data>
 </root>
