+++ tools/SchemaManager/Commands/CommandNames.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-namespace SchemaManager;
+namespace SchemaManager.Commands;
 
 public static class CommandNames
 {
