+++ tools/SchemaManager/Commands/CurrentCommand.cs
@@ -13,8 +13,8 @@ using System.CommandLine.Rendering.Views;
 using System.Threading;
 using System.Threading.Tasks;
 using EnsureThat;
+using Microsoft.Health.SqlServer.Features.Schema.Manager;
 using Microsoft.Health.SqlServer.Features.Schema.Manager.Model;
-using SchemaManager.Core;
 using SchemaManager.Validators;
 
 namespace SchemaManager.Commands;
